backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
python-dotenv==1.0.1
openai==1.54.3
pydantic==2.9.2
